generate GenK instances
