+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/SequenceKTest.kt
@@ -20,8 +20,8 @@ import arrow.core.extensions.sequencek.traverse.traverse
 import arrow.core.extensions.sequencek.unalign.unalign
 import arrow.core.extensions.sequencek.unzip.unzip
 import arrow.test.UnitSpec
-import arrow.test.generators.genK
 import arrow.test.generators.sequenceK
+import arrow.test.generators.sequencek.genK.genK
 import arrow.test.laws.AlignLaws
 import arrow.test.laws.FunctorFilterLaws
 import arrow.test.laws.HashLaws
