+++ modules/core/arrow-test/src/main/kotlin/arrow/test/generators/GenK.kt
@@ -6,15 +6,9 @@ import arrow.core.ForListK
 import arrow.core.ForNonEmptyList
 import arrow.core.ForOption
 import arrow.core.ForSequenceK
-import arrow.core.Id
-import arrow.core.ListK
-import arrow.core.MapK
 import arrow.core.MapKPartialOf
-import arrow.core.NonEmptyList
-import arrow.core.Option
-import arrow.core.SequenceK
-import arrow.core.SortedMapK
 import arrow.core.SortedMapKPartialOf
+import arrow.extension
 import io.kotlintest.properties.Gen
 
 interface GenK<F> {
@@ -24,45 +18,50 @@ interface GenK<F> {
   fun <A> genK(gen: Gen<A>): Gen<Kind<F, A>>
 }
 
-fun Option.Companion.genK() = object : GenK<ForOption> {
+@extension
+interface OptionGenK : GenK<ForOption> {
   override fun <A> genK(gen: Gen<A>): Gen<Kind<ForOption, A>> =
     Gen.option(gen) as Gen<Kind<ForOption, A>>
 }
 
-fun Id.Companion.genK() = object : GenK<ForId> {
+@extension
+interface IdGenK : GenK<ForId> {
   override fun <A> genK(gen: Gen<A>): Gen<Kind<ForId, A>> =
     Gen.id(gen) as Gen<Kind<ForId, A>>
 }
 
-fun ListK.Companion.genK() = object : GenK<ForListK> {
+@extension
+interface ListKGenK : GenK<ForListK> {
   override fun <A> genK(gen: Gen<A>): Gen<Kind<ForListK, A>> =
     Gen.listK(gen) as Gen<Kind<ForListK, A>>
 }
 
-fun NonEmptyList.Companion.genK() = object : GenK<ForNonEmptyList> {
+@extension
+interface NonEmptyListGenK : GenK<ForNonEmptyList> {
   override fun <A> genK(gen: Gen<A>): Gen<Kind<ForNonEmptyList, A>> =
     Gen.nonEmptyList(gen) as Gen<Kind<ForNonEmptyList, A>>
 }
 
-fun SequenceK.Companion.genK() = object : GenK<ForSequenceK> {
+@extension
+interface SequenceKGenK : GenK<ForSequenceK> {
   override fun <A> genK(gen: Gen<A>): Gen<Kind<ForSequenceK, A>> =
     Gen.sequenceK(gen) as Gen<Kind<ForSequenceK, A>>
 }
 
-fun <K> MapK.Companion.genK(kgen: Gen<K>) = object : GenK<MapKPartialOf<K>> {
+@extension
+interface MapKGenK<K> : GenK<MapKPartialOf<K>> {
+
+  fun kGen(): Gen<K>
+
   override fun <A> genK(gen: Gen<A>): Gen<Kind<MapKPartialOf<K>, A>> =
-    Gen.mapK(kgen, gen) as Gen<Kind<MapKPartialOf<K>, A>>
+    Gen.mapK(kGen(), gen) as Gen<Kind<MapKPartialOf<K>, A>>
 }
 
-fun <K : Comparable<K>> SortedMapK.Companion.genK(kgen: Gen<K>) = object : GenK<SortedMapKPartialOf<K>> {
+@extension
+interface SortedMapKGenK<K : Comparable<K>> : GenK<SortedMapKPartialOf<K>> {
+
+  fun kGen(): Gen<K>
+
   override fun <A> genK(gen: Gen<A>): Gen<Kind<SortedMapKPartialOf<K>, A>> =
-    Gen.sortedMapK(kgen, gen) as Gen<Kind<SortedMapKPartialOf<K>, A>>
-}
-
-fun <T> Gen.Companion.id(gen: Gen<T>): Gen<Id<T>> = object : Gen<Id<T>> {
-  override fun constants(): Iterable<Id<T>> =
-    gen.constants().map { Id.just(it) }
-
-  override fun random(): Sequence<Id<T>> =
-    gen.random().map { Id.just(it) }
+    Gen.sortedMapK(kGen(), gen) as Gen<Kind<SortedMapKPartialOf<K>, A>>
 }
